Added dummy listener for base station
This commit adds a subdirectory for the base station project, which is Python code that runs on a raspberry pi. This commit includes a Pipfile for installing dependencies and a README as well as a single Python file test.py. This file just connects to a nrf24l01 and prints what is received.