Added Device ID
Added device IDs for remote sensors so that the base station can differentiate between them. This is implemented as a hardcoded ID built into a the remote sensor firmware. When compiling the firmware, the developer must pass the parameter DEVICE_ID=1 (or a different, unique number) for each sensor. The developer must then re-compile the firmware for each sensor.

This device ID is used as the first byte of the report packet sent back to the base station. The remaining bytes are data (still set to dummy data).

This commit implements this for the firmware as well as changing the dummy base station code to parse out the Device ID and print it.

This commit also updates the README to reflect this build process.