Switched from dummy data to struct with actual reading
Removed dummy data and changed code to send the actual reading from the ADC. For clarity, changed from using an array to using a struct with a uint16_t element for the reading itself. Also changed the corresponding Python code to unpack the binary data using the Python struct library.