Added data replayer
Added a script that imitates the live base station but actually just replays data from a file. This is implemented as run_replay_plotting_base_station.py, which creates two child processes. The first child process is implemented in data_log_player.py, and it loads data from a file and then sends the data out at the recorded times. The second process is the plotting process, the same one as used in the live base station script.