Changed to use pub/sub with ZMQ
This commit changes the scripts from using pipes to sending data over a TCP socket using ZMQ in a publisher/subscriber model. This is convenient because it means we don't have construct complicated arrangements of pipes for each script that is added; the pub/sub model supports multiple publishers and multiple subscribers.

The files that implemented subprocesses that connect to pipes are now replaced with scripts that publish and subscibe to a ZMQ socket:

 - radio_listener.py -> run_radio_relay_publisher.py
 - data_log_player.py -> run_replay_publisher.py
 - data_plotter.py -> run_data_plotter_subscriber.py

The above files can now be run as standalone scripts. The socket information is stored in a new file that they all read, settings.py.

Previously, there existed top-level Python scripts that created the pipe connections between various of the above processes. Although this is no longer necessary since the publishers/subscribers can just be launched directly, shell scripts have been created to replace the two top-level Python scripts that already existed, for convenience:

 - run_plotting_base_station.py -> start_radio_relay_and_plotter.sh
 - run_replay_plotting_base_station.py -> start_replayer_and_plotter.sh

In addition, these two files are unchanged:
 - run_data_logger_base_station.py -- haven't yet implemented logging subscriber
 - run_dummy_base_station.py -- just prints data as it is received, no need for socket communication

Also updated README.